products/serializers.py
# serializers.py

from rest_framework import serializers
from .models import Category, Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'

class CategorySerializer(serializers.ModelSerializer):
    products = ProductSerializer(many=True, read_only=True)  # show products inside category

    class Meta:
        model = Category
        fields = ['id', 'name', 'products']
